pyproject.toml
[project]
name = "bedrock-agent-sdk-sample"
version = "0.1.0"
description = "Claude Agent SDK sample with Bedrock and Langfuse integration"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk>=0.1.0",
    "boto3>=1.35.0",
    "langfuse>=3.10.5",
    "python-dotenv>=1.0.0",
    "anyio>=4.0.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "ipython>=8.0.0",
    "pytest>=8.0.0",
]
evaluation = [
    "deepeval>=1.0.0",
    "langchain-aws>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ipython>=8.0.0",
    "pytest>=8.0.0",
]
